+++ crates/burn-dataset/src/dataset/mod.rs
@@ -3,7 +3,7 @@ mod base;
 mod fake;
 mod in_memory;
 mod iterator;
-mod time_series;
+mod windows;
 #[cfg(any(feature = "sqlite", feature = "sqlite-bundled"))]
 mod sqlite;
